requirements.txt
streamlit==1.28.0
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.0
opencv-python-headless==4.8.1.78
matplotlib==3.7.2
pandas==2.0.3